Angelina/requirements.txt
google-adk
deprecated
fastapi
uvicorn[standard]
python-dotenv
google-genai
pillow